--- a/app/controllers/admin/enterprises_controller.rb
+++ b/app/controllers/admin/enterprises_controller.rb
@@ -154,7 +154,8 @@ module Admin
           Enterprise.where("1=0")
         end
       when :visible
-        OpenFoodNetwork::Permissions.new(spree_current_user).visible_enterprises.ransack(params[:q]).result
+        OpenFoodNetwork::Permissions.new(spree_current_user).visible_enterprises
+          .includes(:shipping_methods, :payment_methods).ransack(params[:q]).result
       else
         # TODO was ordered with is_distributor DESC as well, not sure why or how we want to sort this now
         OpenFoodNetwork::Permissions.new(spree_current_user).