--- a/app/serializers/api/admin/for_order_cycle/enterprise_serializer.rb
+++ b/app/serializers/api/admin/for_order_cycle/enterprise_serializer.rb
@@ -37,7 +37,7 @@ class Api::Admin::ForOrderCycle::EnterpriseSerializer < ActiveModel::Serializer
   end
 
   def products
-    @products ||= products_scope.includes(:supplier, master: [:images])
+    @products ||= products_scope.includes(:variants, :supplier, master: [:images])
   end
 
   def order_cycle