@Matt-Yorkley 2019-05-29
> @Matt-Yorkley is this closing #3858 or is further investigation needed?

It's not closing the spike yet. I'll write up a few things I found and make issues for them if needed.