@lin-d-hop 2019-06-11
Everything is working as expected:
https://docs.google.com/document/d/1RT7GsSCOp8UWFsMznMOz8by6p55o1KmSGBWd5xMc8Cs

Note I did not create enough OCs to have an impact on performance. I did not test if there were performance improvements - only that OC pages all continue to work as expected.

@matt-yorkley - Should we do some benchmarking as this is deployed to prod?