@sauloperez 2019-06-12
> @Matt-Yorkley - Should we do some benchmarking as this is deployed to prod?

Yes! the app behaves quite differently in production compared to development and we need to get accurate estimates as we move forward with the epic.